product serializer

+++ market/serializers.py
@@ -40,16 +40,20 @@ class ProductSerializer(serializers.ModelSerializer):
     def create(self, validated_data):
         request = self.context.get('request')
         if request.user.is_authenticated:
-            info_data = validated_data.pop('info', [])
-            image_data = validated_data.pop('images', [])
-            product = Product.objects.create(owner=request.user, **validated_data)
-            for info in info_data:
-                ProductInfo.objects.create(product=product, **info)
-            for image in image_data:
-                Image.objects.create(product=product, **image)
-            return product
-        else:
-            return Response({'error': 'You are not Authenticated!'}, status=status.HTTP_403_FORBIDDEN)
+            if request.user.is_seller:
+                info_data = validated_data.pop('info', [])
+                image_data = validated_data.pop('images', [])
+                product = Product.objects.create(owner=request.user, **validated_data)
+                if info_data:
+                    for info in info_data:
+                        ProductInfo.objects.create(product=product, **info)
+                if image_data:
+                    for image in image_data:
+                        Image.objects.create(product=product, **image)
+                return product
+            else:
+                raise serializers.ValidationError({'error': 'You are not seller!'})
+        raise serializers.ValidationError({'error': 'You are not Authenticated!'})
 
     def update(self, instance, validated_data):
         request = self.context.get('request')
@@ -58,16 +62,28 @@ class ProductSerializer(serializers.ModelSerializer):
         owner = instance.owner
         if request.user.is_authenticated:
             if request.user == owner:
-                for attr, value in validated_data.items():
-                    setattr(instance, attr, value)
-                instance.save()
-                instance.info.all().delete()
-                for item in info_data:
-                    ProductInfo.objects.create(product=instance, **item)
-                for image in image_data:
-                    Image.objects.create(product=instance, **image)
-                return instance
-            else:
-                return Response({'error': 'You are not owner of this product!'}, status=status.HTTP_403_FORBIDDEN)
-        else:
-            return Response({'error': 'You are not Authenticated!'}, status=status.HTTP_403_FORBIDDEN)
+                if request.user.is_seller:
+                    for attr, value in validated_data.items():
+                        setattr(instance, attr, value)
+                    instance.save()
+                    instance.info.all().delete()
+                    for item in info_data:
+                        ProductInfo.objects.create(product=instance, **item)
+                    for image in image_data:
+                        Image.objects.create(product=instance, **image)
+                    return instance
+                raise serializers.ValidationError({'error': 'You are not seller!'})
+            raise serializers.ValidationError({'error': 'You are not owner of this product!'})
+        raise serializers.ValidationError({'error': 'You are not Authenticated!'})
+
+    def destroy(self, instance):
+        request = self.context.get('request')
+        owner = instance.owner
+        if request.user.is_authenticated:
+            if request.user == owner:
+                if request.user.is_seller:
+                    instance.delete()
+                    return instance
+                raise serializers.ValidationError({"error": "You are not seller!"})
+            raise serializers.ValidationError({"error": "You are not owner of this product!"})
+        raise serializers.ValidationError({'error': 'You are not Authenticated!'})
